More minor improvements and corrections found while playtesting. Thanks sydney_untangle for the help!
When bot is set to never drain, the info about draining results won't be shown.
Fix a typo in Manual of Extract Bot Parts (online -> ingame)
Add an extra node to Chandra, explaining where is everyone.
Rework Open Sesame quest. It will be closed once you show the MS Stock Cert.
-- Please note it tip you are missing something if you haven't met Will Gapes.
Add an node to Karol's dialog asking about bot parts.
Remove backstory of Mike being programmed by Ewald (Act 2 Consistency)
Delay when node66 is shown at Spencer dialog.
Karol now sells Book of Repair Equipment, Barf's drink, Antigrav pod and Pot Helmet.

+++ data/base/dialogs/AfterTakeover.lua
@@ -555,11 +555,14 @@ return {
 				freezetime_seconds = _"second"
 			end
 
-			Npc:says(_"Draining this %s unit will get you %d HP but kill it and take %d %s.",
-				Npc:get_type(),
-				math.ceil(((Npc:get_max_health() - Npc:get_damage())/(difficulty_level()+1))),
-				math.ceil(Aftertakeover_freezetime),
-				freezetime_seconds)
+			if (not (Aftertakeover_drain_attempt == "never")) then
+				--; TRANSLATORS: Example: Draining this 123 unit will get you 5 HP but kill it and take 3 seconds.
+				Npc:says(_"Draining this %s unit will get you %d HP but kill it and take %d %s.",
+					Npc:get_type(),
+					math.ceil(((Npc:get_max_health() - Npc:get_damage())/(difficulty_level()+1))),
+					math.ceil(Aftertakeover_freezetime),
+					freezetime_seconds)
+			end
 
 			if (Aftertakeover_drain_attempt == "ask") then
 				Npc:says(_"Are you sure you want to drain this %s unit?", Npc:get_type())
